pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.9.0",
    "boto3>=1.39.14",
    "fastapi>=0.116.1",
    "pandas>=2.3.1",
    "polars>=1.31.0",
    "python-dotenv>=1.1.1",
    "sagemaker>=2.248.2",
    "torch>=2.7.1",
    "transformers>=4.54.0",
    "uvicorn>=0.35.0",
]
